rename useAlbums hook to useAlbumDocs hook
